GameManager and scene Transition
A change on GameManager.
It is a DonDestroyOnLoad now, So changed the structure on Enemy path. But they are basically same. It is start at Hanu Scene for a test.

Transition trigger, and Transition Animation. The Transition trigger is now in Hanu Scene for a test, and it will do a transition to Lu's test scene. The parameters of the animation is on GameManager